add: debug info and onboard rgb led support

+++ lib/GLOBAL/global.h
@@ -147,9 +147,9 @@ void intToByteArray(int a, byte *byteArray)
   byteArray[0] = a & 0xFF;
 }
 // Function to combine two integers into an unsigned long int
-unsigned long int combineInts(uint32_t int1, uint32_t int2)
+uint32_t combineInts(uint32_t int1, uint32_t int2)
 {
-  return (static_cast<unsigned long int>(int1) << 5) + int2;
+  return (static_cast<uint32_t>(int1) << 5) + int2;
   // unsigned long finalID = (PID << 8) | NodeID;
   // return finalID;
 }
